increased the integration test coverage

+++ src/main/java/frentz/daniel/hardwareservice/service/HardwareService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface HardwareService {
     public Hardware getHardware(long hardwareId);
     public List<Hardware> getAllHardware();
-    Hardware getHardware(String serialNumber, int hardwarePort);
+    Hardware getHardware(String serialNumber, long hardwarePort);
 }
